+++ react/src/index.ts
@@ -26,3 +26,4 @@ export * from './hooks';
 export * from './useBaseURL';
 export * from './useCurrentAgent';
 export * from './useLocalStorage';
+export * from './useServerSearch';