+++ react/src/useBaseURL.ts
@@ -4,36 +4,36 @@ import { useStore } from './hooks';
 import { useLocalStorage } from './useLocalStorage';
 
 /**
- * A hook for using and adjusting the Base URL. Also saves to localStorage. If
+ * A hook for using and adjusting the Server URL. Also saves to localStorage. If
  * the URL is wrong, an error is thrown using the store's handler
  */
-export const useBaseURL = (): [string, (baseURL: string) => void] => {
+export const useBaseURL = (): [string, (serverUrl: string) => void] => {
   // Localstorage for cross-session persistence of JSON object
   const store = useStore();
-  const [baseURLJSON, setBaseURLJSON] = useLocalStorage<string | null>(
-    'baseUrl',
-    store.getBaseUrl(),
+  const [serverUrlJson, setServerUrlJson] = useLocalStorage<string | null>(
+    'serverUrl',
+    store.getServerUrl(),
   );
   const [baseURL, setBaseURL] = useState<string>(window.location.origin);
 
   useEffect(() => {
     if (baseURL !== null) {
-      if (isValidURL(baseURLJSON)) {
-        setBaseURL(baseURLJSON);
+      if (isValidURL(serverUrlJson)) {
+        setBaseURL(serverUrlJson);
       } else {
         store.handleError(
           new Error(
-            `Invalid base URL: ${baseURLJSON}, defaulting to atomicdata.dev`,
+            `Invalid base URL: ${serverUrlJson}, defaulting to atomicdata.dev`,
           ),
         );
         setBaseURL('https://atomicdata.dev');
       }
     }
-  }, [baseURLJSON]);
+  }, [serverUrlJson]);
 
   useEffect(() => {
-    store.setBaseUrl(baseURL);
+    store.setServerUrl(baseURL);
   }, [baseURL]);
 
-  return [baseURL, setBaseURLJSON];
+  return [baseURL, setServerUrlJson];
 };